Move functional API to seaography
